requirements.txt
claude-agent-sdk
anyio
grpcio
grpcio-tools
protobuf
pydantic
aiohttp
swarms
loguru
